Put tests back into their respective files.
